+++ test/order-flows.test.js
@@ -60,6 +60,7 @@ test('H4: reopening a completed order clears stale completion state', () => {
   global.renderAnalytics = () => {}; global.renderDashboard = () => {};
   global.autoExportStatusPage = () => {}; global.autoSendEmailNotification = () => {};
   global.sendTelegramForOrder = () => {}; global.fireWebhook = () => {};
+  global.fireStatusWebhook = () => {};
   global.fireOrderWebhook = () => {};
 
   flows.updateStatus('O1', 'printing');